+++ examples/tutorials/13_mouse_clicking.py
@@ -5,14 +5,8 @@ def main():
     set_render_mode(RenderMode.RENDER_MANUAL)
 
     while is_run():
-        x, y, buttons = get_click()
-        str = "clicked on %d,%d ." % (x, y)
-        if contains_left_button(buttons):
-            str += " left button down"
-        if contains_right_button(buttons):
-            str += " right button down"
-        if contains_mid_button(buttons):
-            str += " mid button down"
+        msg = get_click()
+        str = "clicked on %d,%d ." % (msg.x, msg.y)
         clear_device()
         draw_text(0, 600, str)