+++ examples/tutorials/13_non-blocking_mouse.py
@@ -10,10 +10,10 @@ def main():
         fill_rect(0, 580, 390, 600)
         draw_text(0, 600, "%d,%d" % (x, y))
         if has_mouse_msg():
-            x, y, type, buttons, modifiers = get_mouse_msg()
-            if type == MouseMessageType.PRESS_MESSAGE:
+            msg = get_mouse_msg()
+            if msg.type == MouseMessageType.PRESS_MESSAGE:
                 typestr = "pressed"
-            else:
+            elif msg.type == MouseMessageType.RELEASE_MESSAGE:
                 typestr = "released"
             fill_rect(400, 580, 800, 600)
             draw_text(400, 600, "button %s at %d,%d" % (typestr, x, y))