+++ examples/tutorials/14_change_color.py
@@ -7,8 +7,8 @@ def main():
 
     while is_run():
         if has_mouse_msg():
-            x, y, type, buttons, modifiers = get_mouse_msg()
-            if type == MouseMessageType.PRESS_MESSAGE:
+            msg = get_mouse_msg()
+            if msg.type == MouseMessageType.PRESS_MESSAGE:
                 color = get_color(get_background_color())
                 set_background_color(color)
         delay_fps(60)